% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookFunctions.R
\name{add_sentiments}
\alias{add_sentiments}
\title{add_sentiments}
\usage{
add_sentiments(split_books, sentiments)
}
\arguments{
\item{split_books}{split_books: data frame containing one row per word, space or punctuation mark in each book}

\item{sentiments}{modified sentiments dictionary}
}
\value{
sentiment_books: an extended books dataframe that includes sentiments
}
\description{
joins sentiments given by a dictionary to the books main dataframe
}
\examples{
sentiment_books <- add_sentiments(split_books,sentiments)
}
