% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookFunctions.R
\name{append_sentiments}
\alias{append_sentiments}
\title{append_sentiments}
\usage{
append_sentiments(sentiment_dict)
}
\arguments{
\item{sentiments_dict}{}
}
\value{
sentiments: dataframe with a two column sentiment dictionry
}
\description{
Get sentiments from a given sentiment dictionary
}
\examples{
sentiment_dict <- "nrc"
sentiments <- append_sentiments(sentiment_dict)
}
