% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookFunctions.R
\name{download_books}
\alias{download_books}
\title{downloadBooks}
\usage{
download_books(titles)
}
\arguments{
\item{titles}{Titles of the books to be downloaded}
}
\value{
books: A dataframe with all books stored with one line per row
}
\description{
Download books from Gutenberg project
}
\examples{
titles <- c("The Critique of Pure Reason","Around the World in Eighty Days","Don Quixote")
books <- download_books(titles)
}
