% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookFunctions.R
\name{split_words}
\alias{split_words}
\title{split_words}
\usage{
split_words(books)
}
\arguments{
\item{books}{Data frame of books to be processed}
}
\value{
split_books: data frame containing one row per word, space or punctuation mark in each book
}
\description{
split_words
}
\examples{
split_books <- split_words(books)
}
